project1/requirements.txt
fastapi
uvicorn[standard]
databases[postgresql]
sqlalchemy
psycopg2-binary
pydantic>=2.0
aiohttp
pandas